Oh my god another day wasted trying to make these collisions work properly....
This system is a mess and I definitely dont plan on rebuilding it. My previous two teammates made an absolute mess trying to extend from my oh so simple classes and tied the AI in a knot. No idea currently why the player cannot collide with trails.

Also pushed a minor change to the grid background as it wasn't lined up properly...somehow.