AI now relies on MOVE state for movement.
God damn every one of these changes takes so many tweaks. I definitely didnt factor in the time it would take to deal with Jeff and Tim's messy/broken/unfinished code when I estimate the time it would take to update this project. I should still be able to get the collision avoidance in there before midnight for sure.